Split out questionnaire linking scenario, add tests for all qid type/case combinations
