+++ acceptance_tests/utilities/mappings.py
@@ -520,8 +520,14 @@ QUESTIONNAIRE_TYPE_TO_FORM_TYPE = {
     "01": 'H',
     "02": 'H',
     "04": 'H',
+    "03": 'H',
+    "11": 'cont',
+    "12": 'cont',
+    "13": 'cont',
+    "14": 'cont',
     "21": 'I',
     "22": 'I',
+    "23": 'I',
     "24": 'I',
     "31": 'C',
     "32": 'C',
